File attach on objects

+++ app.py
@@ -140,6 +140,9 @@ def load_object(ID,mode):
 	query = DB.get_object_parameters(id=ID, projectId=projectSession())
 	areaInObject = DB.getObjectArea(objectId=ID)
 	areas = DB.getAreas(projectId=projectSession())
+	files = DB.getObjectFiles(obId=ID)
+	if files == []:
+		files="empty"
 	temp=[]
 	boolen='false'
 	for k in areas:
@@ -152,9 +155,9 @@ def load_object(ID,mode):
 			temp.append([k,'checked'])
 			boolen='false'
 	if mode == "loadObject":
-		return render_template("object.html", loadObject=query,areas=temp,count=len(areas))
+		return render_template("object.html", loadObject=query,areas=temp,count=len(areas),files=files)
 	elif mode == "editObject":
-		return render_template("object.html", editObject=query,areas=temp,count=len(areas))
+		return render_template("object.html", editObject=query,areas=temp,count=len(areas),files=files)
 
 @app.route('/save_object', methods=['POST'])	
 def save_object():
@@ -772,15 +775,17 @@ def allowed_file(filename):
     return '.' in filename and \
            filename.rsplit('.', 1)[1] in ALLOWED_EXTENSIONS
 
-@app.route('/upload_file_test/<int:stepexeId>', methods=['GET', 'POST'])
-def upload_file_test(stepexeId):
-	UPLOAD_FOLDER = './uploads/test/'
+@app.route('/upload_file_test/<int:Id>/<mode>', methods=['GET', 'POST'])
+def upload_file_test(Id,mode):
+	if mode == "test":
+		UPLOAD_FOLDER = './uploads/test/'
+	if mode == "object":
+		UPLOAD_FOLDER = './uploads/object/'
 	app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
 	if request.method == 'POST':
 		if request.form['name'] == '':
 			return "No selected file"
 		if allowed_file(request.form['name']):
-			print(request.form['name'])
 			if request.form['name'].rsplit('.', 1)[1] in PIC_EXTENSIONS:
 				file = open(os.path.join(app.config['UPLOAD_FOLDER'], request.form['name']),'wb')
 				file.write(base64.b64decode(request.form['context'].split(',')[1]))
@@ -788,8 +793,12 @@ def upload_file_test(stepexeId):
 				file = open(os.path.join(app.config['UPLOAD_FOLDER'], request.form['name']),'w')
 				file.write(request.form['context'])
 			file.close()
-			result=UP.saveTestFile(stepexeId=stepexeId,url=os.path.join(app.config['UPLOAD_FOLDER'], request.form['name']),filename=request.form['name'],extension=request.form['name'].rsplit('.', 1)[1])
-			return render_template("upload.html", addPlusToTest=result)
+			if mode == "test":
+				result=UP.saveTestFile(stepexeId=Id,url=os.path.join(app.config['UPLOAD_FOLDER'], request.form['name']),filename=request.form['name'],extension=request.form['name'].rsplit('.', 1)[1])
+				return render_template("upload.html", addPlusToTest=result)
+			if mode == "object":
+				result=UP.saveObjectFile(objectId=Id,url=os.path.join(app.config['UPLOAD_FOLDER'], request.form['name']),filename=request.form['name'],extension=request.form['name'].rsplit('.', 1)[1])
+				return render_template("upload.html", objectFile=result)
 	return "error"
 
 @app.route('/fileProperty/<int:exeId>/<int:stepId>', methods=['GET'])	
@@ -800,11 +809,40 @@ def fileProperty(exeId,stepId):
 		files="empty"
 	return render_template('upload.html', testUpload=files, stepId=stepId,stepexeId=stepexeId)
 
-@app.route('/deleteFileTest/<int:fileId>', methods=['GET'])	
-def deleteFileTest(fileId):
-	file=UP.getFileURL(fileId=fileId)
-	os.remove(file)
-	return UP.deleteFileTest(fileId=fileId)
+@app.route('/upload_file_update/<int:id>/<mode>', methods=['POST'])	
+def upload_file_update(id,mode):
+	if mode == "object":
+		UPLOAD_FOLDER = './uploads/object/'
+	app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
+	if request.method == 'POST':
+		if request.form['name'] == '':
+			return "No selected file"
+		if allowed_file(request.form['name']):
+			if mode == "object":
+				result=DB.checkFileInObjects(objectId=id,filename=request.form['name'])
+				if result == None:
+					if request.form['name'].rsplit('.', 1)[1] in PIC_EXTENSIONS:
+						file = open(os.path.join(app.config['UPLOAD_FOLDER'], request.form['name']),'wb')
+						file.write(base64.b64decode(request.form['context'].split(',')[1]))
+					else:
+						file = open(os.path.join(app.config['UPLOAD_FOLDER'], request.form['name']),'w')
+						file.write(request.form['context'])
+					file.close()
+					result=UP.saveObjectFile(objectId=id,url=os.path.join(app.config['UPLOAD_FOLDER'], request.form['name']),filename=request.form['name'],extension=request.form['name'].rsplit('.', 1)[1])
+					return render_template('upload.html', objectFile=result)
+		return "ok"
+	return "error"
+	
+@app.route('/deleteFiles/<int:fileId>/<mode>', methods=['GET'])	
+def deleteFiles(fileId,mode):
+	if mode == "test":
+		file=UP.getTestFileURL(fileId=fileId)
+		os.remove(file)
+		return UP.deleteFileTest(fileId=fileId)
+	if mode == "object":
+		file=UP.getObjectFileURL(fileId=fileId)
+		os.remove(file)
+		return UP.deleteFileObject(fileId=fileId)
 	
 # set the secret key.  keep this really secret:
 app.secret_key = os.urandom(24) #'A0Zr98j/3yX R~XHH!jmN]LWX/,?RT'
